+++ tests/JsonSerializerTest.php
@@ -2,10 +2,11 @@
 
 namespace Zumba\JsonSerializer\Test;
 
+use Zumba\JsonSerializer\ClosureSerializer;
 use Zumba\JsonSerializer\JsonSerializer;
 use Zumba\JsonSerializer\Exception\JsonSerializerException;
 use stdClass;
-use SuperClosure\Serializer as ClosureSerializer;
+use SuperClosure\Serializer as SuperClosureSerializer;
 use PHPUnit\Framework\TestCase;
 use ReflectionProperty;
 
@@ -246,7 +247,7 @@ class JsonSerializerTest extends TestCase
      *
      * @return void
      */
-    public function testSerializeEnums() 
+    public function testSerializeEnums()
     {
         if (PHP_VERSION_ID < 80100) {
             $this->markTestSkipped("Enums are only available since PHP 8.1");
@@ -266,7 +267,7 @@ class JsonSerializerTest extends TestCase
      *
      * @return void
      */
-    public function testUnserializeEnums() 
+    public function testUnserializeEnums()
     {
         if (PHP_VERSION_ID < 80100) {
             $this->markTestSkipped("Enums are only available since PHP 8.1");
@@ -389,13 +390,13 @@ class JsonSerializerTest extends TestCase
      *
      * @return void
      */
-    public function testSerializationOfClosure()
+    public function testSerializationOfClosureWithSuperClosureOnConstructor()
     {
         if (!class_exists('SuperClosure\Serializer')) {
             $this->markTestSkipped('SuperClosure is not installed.');
         }
 
-        $closureSerializer = new ClosureSerializer();
+        $closureSerializer = new SuperClosureSerializer();
         $serializer = new JsonSerializer($closureSerializer);
         $serialized = $serializer->serialize(
             array(
@@ -414,6 +415,111 @@ class JsonSerializerTest extends TestCase
     }
 
     /**
+     * Test the serialization of closures providing closure serializer
+     *
+     * @return void
+     */
+    public function testSerializationOfClosureWithSuperClosureOnManager()
+    {
+        if (!class_exists('SuperClosure\Serializer')) {
+            $this->markTestSkipped('SuperClosure is not installed.');
+        }
+
+        $closureSerializer = new SuperClosureSerializer();
+        $serializer = new JsonSerializer();
+        $serializer->addClosureSerializer(new ClosureSerializer\SuperClosureSerializer($closureSerializer));
+        $serialized = $serializer->serialize(
+            array(
+            'func' => function () {
+                return 'it works';
+            },
+            'nice' => true
+            )
+        );
+
+        $unserialized = $serializer->unserialize($serialized);
+        $this->assertTrue(is_array($unserialized));
+        $this->assertTrue($unserialized['nice']);
+        $this->assertInstanceOf('Closure', $unserialized['func']);
+        $this->assertSame('it works', $unserialized['func']());
+    }
+
+    /**
+     * Test the serialization of closures providing closure serializer
+     *
+     * @return void
+     */
+    public function testSerializationOfClosureWitOpisClosure()
+    {
+        if (!class_exists('Opis\Closure\SerializableClosure')) {
+            $this->markTestSkipped('OpisClosure is not installed.');
+        }
+
+        $serializer = new JsonSerializer();
+        $serializer->addClosureSerializer(new ClosureSerializer\OpisClosureSerializer());
+        $serialized = $serializer->serialize(
+            array(
+            'func' => function () {
+                return 'it works';
+            },
+            'nice' => true
+            )
+        );
+
+        $unserialized = $serializer->unserialize($serialized);
+        $this->assertTrue(is_array($unserialized));
+        $this->assertTrue($unserialized['nice']);
+        $this->assertInstanceOf('Closure', $unserialized['func']);
+        $this->assertSame('it works', $unserialized['func']());
+    }
+
+    /**
+     * Test the serialization of closures providing closure serializer
+     *
+     * @return void
+     */
+    public function testSerializationOfClosureWitMultipleClosures()
+    {
+        if (!class_exists('SuperClosure\Serializer')) {
+            $this->markTestSkipped('SuperClosure is not installed.');
+        }
+        if (!class_exists('Opis\Closure\SerializableClosure')) {
+            $this->markTestSkipped('OpisClosure is not installed.');
+        }
+
+        $closureSerializer = new SuperClosureSerializer();
+        $serializer = new JsonSerializer();
+        $serializer->addClosureSerializer(new ClosureSerializer\SuperClosureSerializer($closureSerializer));
+
+        $serializeData = array(
+            'func' => function () {
+                return 'it works';
+            },
+            'nice' => true
+        );
+
+        // Make sure it was serialized with SuperClosure
+        $serialized = $serializer->serialize($serializeData);
+        echo $serialized;
+        $this->assertGreaterThanOrEqual(0, strpos($serialized, 'SuperClosure'));
+        $this->assertFalse(strpos($serialized, 'OpisClosure'));
+
+        // Test adding a new preferred closure serializer
+        $serializer->addClosureSerializer(new ClosureSerializer\OpisClosureSerializer());
+
+        $unserialized = $serializer->unserialize($serialized);
+        $this->assertTrue(is_array($unserialized));
+        $this->assertTrue($unserialized['nice']);
+        $this->assertInstanceOf('Closure', $unserialized['func']);
+        $this->assertSame('it works', $unserialized['func']());
+
+        // Serialize again with the new preferred closure serializer
+        $serialized = $serializer->serialize($serializeData);
+        $this->assertFalse(strpos($serialized, 'SuperClosure'));
+        $this->assertGreaterThanOrEqual(0, strpos($serialized, 'OpisClosure'));
+    }
+
+    /**
      * Test the unserialization of closures without providing closure serializer
      *
      * @return void
@@ -424,7 +530,7 @@ class JsonSerializerTest extends TestCase
             $this->markTestSkipped('SuperClosure is not installed.');
         }
 
-        $closureSerializer = new ClosureSerializer();
+        $closureSerializer = new SuperClosureSerializer();
         $serializer = new JsonSerializer($closureSerializer);
         $serialized = $serializer->serialize(
             array(