+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Support of UnitEnum and BackedEnum deserialization (PHP > 8.1)
+### Fixed
+- Fixed deprecated with DateTimeImmutable deserialization with PHP 8.2
 
 ## [3.0.0] - 2020-07-20
 ### Fixed